refactor: interface, impl 구분
