+++ src/main/java/kitchenpos/menu/domain/service/MenuGroupService.java
@@ -1,36 +1,9 @@
 package kitchenpos.menu.domain.service;
 
 import java.util.List;
-import java.util.Objects;
-import java.util.UUID;
 import kitchenpos.menu.domain.entity.MenuGroup;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-public class MenuGroupService {
-
-    private final MenuGroupRepository menuGroupRepository;
-
-    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
-        this.menuGroupRepository = menuGroupRepository;
-    }
-
-    @Transactional
-    public MenuGroup create(final MenuGroup request) {
-        final String name = request.getName();
-        if (Objects.isNull(name) || name.isEmpty()) {
-            throw new IllegalArgumentException();
-        }
-        final MenuGroup menuGroup = new MenuGroup();
-        menuGroup.setId(UUID.randomUUID());
-        menuGroup.setName(name);
-        return menuGroupRepository.save(menuGroup);
-    }
-
-    @Transactional(readOnly = true)
-    public List<MenuGroup> findAll() {
-        return menuGroupRepository.findAll();
-    }
+public interface MenuGroupService {
+    MenuGroup create(final MenuGroup request);
+    List<MenuGroup> findAll();
 }
